+++ modules/promotions/widgets/atomic-form-widget-promotion.php
@@ -83,7 +83,7 @@ class Atomic_Form_Widget_Promotion {
 			'title' => __( 'Atomic form', 'elementor' ),
 			'content' => __( 'Design fully customized forms to capture leads without compromising on style.', 'elementor' ),
 			'ctaText' => __( 'Upgrade now', 'elementor' ),
-			'animation' => 'atomic-form',
+			'animation' => 'atomic-form-animation',
 			'widgetCtaUrl' => 'https://go.elementor.com/go-pro-atomic-form-modal/',
 			'sectionCtaUrl' => 'https://go.elementor.com/go-pro-atomic-form-section/',
 		];
